add manual processing

+++ Collocation/collocation.py
@@ -10,11 +10,15 @@ import csv
 nlp = spacy.load("en_core_sci_lg")
 
 input_spec="concordance-download_1985.txt"
+output_specs="maria_wahatever.csv"
+
+no_option=True
 print(f"{sys.argv=}")
 for i in range(len(sys.argv)):
     if sys.argv[i]=="-i":
         if i+1<len(sys.argv):
             input_spec=sys.argv[i+1]
+            no_option=False
             break
 print(f"{input_spec=}")
 
@@ -26,11 +30,12 @@ if input_spec=="all":
         input_files.append(item)
 else:
     input_files.append(input_spec)
-    for item in os.listdir("."):
-        if not item.startswith("concordance-download_"):continue
-        if not item.endswith(".txt"):continue
-        if item==input_spec:continue
-        os.unlink(item)
+    if no_option==False:
+        for item in os.listdir("."):
+            if not item.startswith("concordance-download_"):continue
+            if not item.endswith(".txt"):continue
+            if item==input_spec:continue
+            os.unlink(item)
 
 print(f"{input_files=}")
 
@@ -48,6 +53,9 @@ for input_file in input_files:
 
     # Prepare the CSV file
     output_file=input_file.replace("concordance-download_","collocation_novels_").replace(".txt","_sci.csv")
+    #Maria's choice of output
+    if no_option==True:
+        output_file=output_specs
     print(f"Processing {output_file=} ...")
     with open(output_file, mode="w", newline="") as csv_file:
         fieldnames = ["Original Text", "Relevant Noun", "Relevant Lemma"]
